==> 1. Exploring the data .ipynb <==
# 1. Exploring the data 

import numpy as np 
import pandas as pd
import re

def open_file(path):
    myfile = open(path)
    print(myfile.read())

## Cleaning Episode VI

episodeVI = pd.read_csv(r"..\Project_4_Sentiment-Analysis-API\raw_data\SW_EpisodeVI.txt", 
                        delim_whitespace=True, names=["index","character","dialogue", "other", "other2", "other3", "other4", 
                                                      "other5", "other6", "other7"] ,header = None)

# creating a dataframe with the txt file, for some reason I had to create
# additional columns to store some values 

episodeVI[0:10] #displaying the data - I can see there is really Nan for most of them

episodeVI["other2"].unique() #Ok now I know I have in column other 2 I have

episodeVI[episodeVI["other"] == 'hardly'] 
#line 284 is the line that is crashing the entire ds

episodeVI["dialogue"][284] = f' {episodeVI["dialogue"][284]} {episodeVI["other"][284]} {episodeVI["other2"][284]} {episodeVI["other3"][284]} {episodeVI["other4"][284]} {episodeVI["other5"][284]} {episodeVI["other6"][284]}'
episodeVI["dialogue"][284]

episodeVI["other"].unique() #Ok now I know I have in column other 2 I have

episodeVI[episodeVI["other"] == 'welcome.\\""'] 


episodeVI["dialogue"][190] = f'{episodeVI["dialogue"][190]} {episodeVI["other"][190]}'
episodeVI["dialogue"][190]

def column_cleaning(episode,column_name):
    episode.drop([column_name], axis=1,inplace=True)
    return episode

episodeVI.head()

column_cleaning(episodeVI,'other')
column_cleaning(episodeVI,'other2')
column_cleaning(episodeVI,'other3')
column_cleaning(episodeVI,'other4')
column_cleaning(episodeVI,'other5')
column_cleaning(episodeVI,'other6')
column_cleaning(episodeVI,'other7')

episodeVI[0:10]

def final_df(episode, name):
    episode = episode.drop([0])
    episode.drop(['index'], axis = 1, inplace = True) 
    episode['episode'] = name
    return episode


episodeVI = final_df(episodeVI, 'episode VI')
episodeVI[0:10]

## Cleaning Episode IV

episodeIV = pd.read_csv(r"..\Project_4_Sentiment-Analysis-API\raw_data\SW_EpisodeIV.txt", 
                        delim_whitespace=True, names=["index","character","dialogue","other", "other2", "other3", "other4", "other5", "other6"] ,header = None)

episodeIV[0:10]

episodeIV["other"].unique() #finding unique value

episodeIV[episodeIV["other"] == 'found'] #in which line there is a unique value


episodeIV["dialogue"][554] = f'{episodeIV["dialogue"][554]} {episodeIV["other"][554]} {episodeIV["other2"][554]} {episodeIV["other3"][554]} {episodeIV["other4"][554]} {episodeIV["other5"][554]} {episodeIV["other"][554]} {episodeIV["other6"][554]}'
episodeIV["dialogue"][554]

column_cleaning(episodeIV,'other')
column_cleaning(episodeIV,'other2')
column_cleaning(episodeIV,'other3')
column_cleaning(episodeIV,'other4')
column_cleaning(episodeIV,'other5')
column_cleaning(episodeIV,'other6')

episodeIV = final_df(episodeIV, 'episode IV')
episodeIV[0:10]

## Cleaning Episode V

episodeV = pd.read_csv(r"..\Project_4_Sentiment-Analysis-API\raw_data\SW_EpisodeV.txt", 
                        delim_whitespace=True, names=["index","character","dialogue", "other"] ,header = None)

episodeV[0:10]

episodeV["other"].unique() #finding unique value

episodeV[episodeV["other"] == 'knows\\"?"']

episodeV["dialogue"][53] = f'{episodeV["dialogue"][53]} {episodeV["other"][53]}'
episodeV["dialogue"][53]

column_cleaning(episodeV,'other')

episodeV =final_df(episodeV, 'episode V')
episodeV.head()

### Merging the 3 episode + Creating csv files

merged_data_final = pd.concat([episodeIV, episodeV, episodeVI])
merged_data_final

merged_data_final['dialogue'] = merged_data_final['dialogue'].str.replace('[^\w\s#@/:%.,_-]', '', regex=True, flags=re.UNICODE)

merged_data_final[0:10]

merged_data_final.to_csv("df_all_episodes.csv", index=False, encoding='utf-8')
