==> 2. Creating my own Database with MySQL.ipynb <==
## 2. Creating my own Database with MySQL

import pandas as pd
import sqlalchemy as alch
from getpass import getpass
import re
from dotenv import load_dotenv
import sqlalchemy as sqlalc

df_episode_final = pd.read_csv(r"..\Project_4_Sentiment-Analysis-API\cleaned_data\df_all_episodes.csv")

df_episode_final.head()

### Connect to SQL

# password = getpass("Inseet your password here: ")

# dbName = "starwars"

# connectionData = f"mysql+pymysql://root:{password}@localhost/{dbName}"

# engine = alch.create_engine(connectionData)

# list(engine.execute("SELECT * FROM dialogue"))

```mysql
-- MySQL Workbench Forward Engineering

SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;
SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;
SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';

-- -----------------------------------------------------
-- Schema starwars
-- -----------------------------------------------------

-- -----------------------------------------------------
-- Schema starwars
-- -----------------------------------------------------
CREATE SCHEMA IF NOT EXISTS `starwars` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci ;
-- -----------------------------------------------------
-- Schema star wars
-- -----------------------------------------------------
-- -----------------------------------------------------
-- Schema starwars
-- -----------------------------------------------------

-- -----------------------------------------------------
-- Schema starwars
-- -----------------------------------------------------
CREATE SCHEMA IF NOT EXISTS `starwars` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci ;
USE `starwars` ;

-- -----------------------------------------------------
-- Table `starwars`.`dialogue`
-- -----------------------------------------------------
CREATE TABLE IF NOT EXISTS `starwars`.`dialogue` (
  `id` INT NOT NULL AUTO_INCREMENT,
  `character` VARCHAR(45) NULL,
  `dialogue` TEXT NULL,
  `episode` VARCHAR(45) NULL,
  PRIMARY KEY (`id`))
ENGINE = InnoDB;

USE `starwars` ;

SET SQL_MODE=@OLD_SQL_MODE;
SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
```

### Checking function

# def checkInfo(table, str2):
    
    #if table == "dialogue":
        #query = list(engine.execute(f"SELECT text FROM dialogue WHERE text = '{str2}';"))
        #if len(query) > 0:
            #return True
        #else:
            #return False

#def insertDialogue(str1, str2, str3):
    #if checkInfo("dialogue", str2):
        #return "It already exists"
    #else:
        #engine.execute(f"INSERT INTO starwars(character, text, episode) VALUE ('{str1}', '{str2}', '{str3}');")

#for index, row in df_episode_final.iterrows():
    #try:
        #insertDialogue(row["character"], row["dialogue"], row["episode"])
    #except (ValueError, TypeError):
        #pass

### New Solution 

I used the below code to upload the data and create the database in MySQL

drop database if exists starwars; 
create database starwars;
use starwars;
drop table if exists dialogue; 

create table dialogue
(character_ varchar(40), 
text_ text, 
episode varchar(40)
); 

LOAD DATA INFILE 'C:/ProgramData/MySQL/MySQL Server 8.0/Uploads/df_all_episodes.csv'
	INTO TABLE dialogue
		FIELDS TERMINATED BY ','
        OPTIONALLY ENCLOSED BY '"'
        LINES TERMINATED BY '\r\n'
        IGNORE 1 ROWS;

select count(*) from dialogue;

SELECT character_ from dialogue GROUP BY character_
