==> 4. Creating an API .ipynb <==
## 4. Creating an API

from flask import Flask
import random
from pymongo import MongoClient
import pymongo
import tools
import tools.sqltools as sql
from flask import jsonify
from os import name
from flask import Flask, request, jsonify
import markdown.extensions.fenced_code
import pandas as pd
import json
import random
import googletrans
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from nltk.corpus import stopwords
import statistics as st

password = getpass("Introduce your password: ")
dbName = "starwars"
connectionData=f"mysql+pymysql://root:{password}@localhost/{dbName}"
engine = alch.create_engine(connectionData)

query = "SELECT * FROM dialogue"
df = pd.read_sql_query(query, engine)
df

import requests

### GET

response = requests.get("http://127.0.0.1:5000/line").json()


app = Flask(__name__)

@app.route("/")
def greeting ():
    return f"How are you doing"


@app.route("/random-number")
def random_number ():
    return str(random.choice(range(0,11)))

@app.route("/campus/<location>")
def campus_location (location):
    if location == "bcn":
        return "Carrer Pamplona 96"
    elif location == "mad":
        return "Paseo de la chopera, 14"
    else: 
        return "No campus here :/"
      
if __name__ == '__main__':
    app.run()

what can things you can answer 
most frequent words
most frequeent characters 
and then run sentimental anaylsis 
exemple of get/post query
